Game script and CLI Output
Start of game script. Run through command line. Welcomes user, then asks
to create a game, asks name, and to start.

On game start, output formatter takes the field in play and converts to
a visual representation of the field as the cli permits. Showing
features in their shape, fill, color, and count. It will also output
each card as in a 4x3 table (5x3 if there are more cards).